requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cleanup==7.0.0
django-crontab==0.7.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.6
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
